+++ src/feature/home/types/shortPostData.ts
@@ -1,0 +1,6 @@
+export interface ShortPostData {
+  id: string;
+  title: string;
+  updated_at: string;
+  tagname: string | null;
+}
