+++ tailwind.config.ts
@@ -16,13 +16,13 @@ const config: Config = {
         },
       },
       fontSize: {
-        xs: ['12px', { lineHeight: '1' }],
-        sm: ['14px', { lineHeight: '1' }],
-        base: ['16px', { lineHeight: '1.5' }],
-        lg: ['18px', { lineHeight: '1.5' }],
-        xl: ['20px', { lineHeight: '1', fontWeight: '700' }],
-        '2xl': ['24px', { lineHeight: '1' }],
-        '3xl': ['30px', { lineHeight: '1' }],
+        xs: '12px',
+        sm: '14px',
+        base: '16px',
+        lg: '18px',
+        xl: '20px',
+        '2xl': '24px',
+        '3xl': '30px',
       },
       height: {
         1: '4px',
